Import Leaves

Imports leaves created by the QSM foliage and needles naive insertion (QSM-FaNNI) algorithm. The leaves are stored as a triangular mesh, compatible with both rTwig and RGL functions.

Usage

import_leaves(filename, format = "obj", simplify = FALSE)

Arguments

filename

a QSM-FaNNI .obj file

format

.obj file format. Valid formats include obj and obj_ext. Defaults to obj.

simplify

simplify the mesh by removing duplicate vertices. This can reduce memory usage at the cost of import speed. Defaults to FALSE.

Value

rgl::mesh3d object

References

FaNNIrTwig

Examples

if (FALSE) { # \dontrun{

## Import FaNNI leaves from the standard obj format
file <- "path_to_leaves.obj"
qsm <- import_leaves(file, format = "obj")

## Import FaNNI leaves from the extended obj format
file <- "path_to_leaves_extended.obj"
qsm <- import_leaves(file, format = "obj_ext")
} # }