Tree Metrics

Calculates tree metrics from a QSM

Usage

tree_metrics(cylinder)

Arguments

cylinder

QSM cylinder data frame

Value

Returns a list of tree metric data frames and synthetic point cloud

Details

Calculates detailed tree metrics from a QSM. The outputs include all of the standard outputs from TreeQSM, and also additional variables, including, but not limited to, growth length, reverse branch order, branch segment or node relationships, and distances from twigs and the base of the tree, across various distribution metrics. Also included is a simulated point cloud of the tree, based on the QSM cylinder radii. When corrected with Real Twig, this allow for the testing and validation of point cloud diameter overestimation throughout the tree.

References

Raumonen P, Kaasalainen M, Åkerblom M, Kaasalainen S, Kaartinen H, Vastaranta M, Holopainen M, Disney M, Lewis P (2013). “Fast Automatic Precision Tree Models from Terrestrial Laser Scanner Data.” Remote Sensing, 5(2), 491--520. doi:10.3390/rs5020491 .

Hackenberg J, Spiecker H, Calders K, Disney M, Raumonen P (2015). “SimpleTree —An Efficient Open Source Tool to Build Tree Models from TLS Clouds.” Forests, 6(11), 4245--4294. doi:10.3390/f6114245 .

Hackenberg J, Bontemps J (2023). “Improving quantitative structure models with filters based on allometric scaling theory.” Applied Geomatics, 15. doi:10.1007/s12518-023-00537-4 .

Yang W, Wilkes P, Vicari MB, Hand K, Calders K, Disney M (2024). “Treegraph: tree architecture from terrestrial laser scanning point clouds.” Remote Sensing in Ecology and Conservation. ISSN 2056-3485, doi:10.1002/rse2.399 .

Examples

## TreeQSM Processing Chain
file <- system.file("extdata/QSM.mat", package = "rTwig")
cylinder <- import_qsm(file)$cylinder
#> Importing TreeQSM .mat
cylinder <- update_cylinders(cylinder)
#> Updating Cylinder Ordering
#> Calculating Total Children
#> Building Cylinder Network
#> Calculating Growth Length
#> Calculating Reverse Branch Order
#> Calculating Branch Segments
#> Calculating Distance From Base
#> Calculating Average Distance To Twigs
metrics <- tree_metrics(cylinder)
#> Calculating Branch Metrics
#> Calculating Tree Metrics
#> Generating Point Cloud
#> Calculating Crown Metrics
#> Calculating Tree Distributions
#> Calculating Branch Distributions
#> Calculating Stem Taper
#> Calculating Spreads
str(metrics)
#> List of 17
#>  $ branch                   :Classes ‘grouped_tt’, ‘tidytable’, ‘data.table’ and 'data.frame':	82 obs. of  19 variables:
#>   ..$ branch          : int [1:82] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ parent_branch   : int [1:82] 0 1 1 1 1 1 1 1 1 1 ...
#>   ..$ branch_order    : int [1:82] 0 1 1 1 1 1 1 1 1 1 ...
#>   ..$ reverse_order   : num [1:82] 20 4 1 6 16 1 1 1 7 1 ...
#>   ..$ diameter_base_cm: num [1:82] 9.29 3.35 1.11 3.37 4.75 ...
#>   ..$ diameter_mid_cm : num [1:82] 2.47 1.04 0.587 1.128 1.374 ...
#>   ..$ diameter_tip_cm : num [1:82] 1.963 1.121 0.882 2.925 0.856 ...
#>   ..$ volume_m3       : num [1:82] 1.08e-02 5.18e-04 1.38e-05 9.04e-04 2.10e-03 ...
#>   ..$ area_m2         : num [1:82] 0.64394 0.08328 0.00503 0.13603 0.23585 ...
#>   ..$ length_m        : num [1:82] 3.461 1.177 0.147 1.708 2.399 ...
#>   ..$ height_m        : num [1:82] 0 1.3 1.67 1.75 1.82 ...
#>   ..$ angle_deg       : num [1:82] 0 36.3 69 49.6 53.7 ...
#>   ..$ azimuth_deg     : num [1:82] 83.77 4.72 -134.61 31.17 136.74 ...
#>   ..$ zenith_deg      : num [1:82] 5.26 42.37 70.3 63.45 27.38 ...
#>   ..$ growth_length   : num [1:82] 32.8 2.01 0.147 4.985 12.793 ...
#>   ..$ nodes           : int [1:82] 14 4 1 6 15 1 1 1 7 1 ...
#>   ..$ children        : num [1:82] 14 4 1 6 15 1 1 1 7 1 ...
#>   ..$ base_distance_m : num [1:82] 0.0939 3.393 4.2416 4.164 3.3096 ...
#>   ..$ twig_distance_m : num [1:82] 3.405 0.871 0.147 1.397 1.75 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..- attr(*, "groups")= chr "branch"
#>  $ cloud                    : num [1:10644, 1:3] 0.758 0.77 0.782 0.793 0.803 ...
#>  $ tree                     :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	1 obs. of  27 variables:
#>   ..$ tree_volume_m3         : num 0.0222
#>   ..$ stem_volume_m3         : num 0.0108
#>   ..$ branch_volume_m3       : num 0.0114
#>   ..$ tree_height_m          : num 3.69
#>   ..$ tree_length_m          : num 32.8
#>   ..$ stem_length_m          : num 3.46
#>   ..$ branch_length_m        : num 29.3
#>   ..$ branches               : int 82
#>   ..$ max_branch_order       : int 5
#>   ..$ max_reverse_order      : num 20
#>   ..$ tree_area_m2           : num 2.53
#>   ..$ stem_area_m2           : num 0.644
#>   ..$ branch_area_m2         : num 1.88
#>   ..$ dbh_qsm_cm             : num 7.18
#>   ..$ dbh_raw_cm             : num 7.18
#>   ..$ crown_diameter_mean_m  : num 1.53
#>   ..$ crown_diameter_max_m   : num 2.31
#>   ..$ crown_diameter_min_m   : num 0.881
#>   ..$ crown_projected_area_m2: num 3.34
#>   ..$ crown_base_height_m    : num 1.3
#>   ..$ crown_length_m         : num 2.39
#>   ..$ crown_ratio            : num 0.647
#>   ..$ crown_area_m2          : num 13.6
#>   ..$ crown_volume_m3        : num 0.00366
#>   ..$ start_x                : num 0.00995
#>   ..$ start_y                : num 0.0912
#>   ..$ start_z                : num 0.996
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_height_dist         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	4 obs. of  4 variables:
#>   ..$ height_class_m: int [1:4] 1 2 3 4
#>   ..$ volume_m3     : num [1:4] 0.00542 0.00551 0.00729 0.00393
#>   ..$ area_m2       : num [1:4] 0.261 0.475 1.077 0.714
#>   ..$ length_m      : num [1:4] 1 4.54 15.17 12.09
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_diameter_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	11 obs. of  4 variables:
#>   ..$ diameter_class_cm: num [1:11] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ volume_m3        : num [1:11] 7.41e-05 3.08e-03 3.34e-03 4.19e-03 1.86e-03 ...
#>   ..$ area_m2          : num [1:11] 0.0349 0.7716 0.5698 0.4906 0.1611 ...
#>   ..$ length_m         : num [1:11] 1.34 15.82 7.83 4.6 1.11 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_zenith_dist         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	15 obs. of  4 variables:
#>   ..$ zenith_class_deg: num [1:15] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3       : num [1:15] 0.00947 0.00222 0.0026 0.00141 0.00159 ...
#>   ..$ area_m2         : num [1:15] 0.534 0.255 0.315 0.23 0.278 ...
#>   ..$ length_m        : num [1:15] 2.8 2.86 3.94 3.56 4.54 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_azimuth_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	36 obs. of  4 variables:
#>   ..$ azimuth_class_deg: num [1:36] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3        : num [1:36] 0.000404 0.000139 0.000333 0.000128 0.000152 ...
#>   ..$ area_m2          : num [1:36] 0.0702 0.0308 0.0371 0.0298 0.0255 ...
#>   ..$ length_m         : num [1:36] 1.099 0.582 0.597 0.636 0.46 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_diameter_dist     :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	5 obs. of  9 variables:
#>   ..$ diameter_class_cm: num [1:5] 1 2 3 4 5
#>   ..$ volume_m3        : num [1:5] 0.00 2.19e-05 6.72e-04 2.01e-03 1.96e-03
#>   ..$ volume_1_m3      : num [1:5] 0.00 0.00 5.82e-05 2.44e-04 6.75e-04
#>   ..$ area_m2          : num [1:5] 0 0.0109 0.1743 0.3908 0.4042
#>   ..$ area_1_m2        : num [1:5] 0 0 0.0187 0.058 0.1312
#>   ..$ length_m         : num [1:5] 0 0.44 4.12 7.11 6.89
#>   ..$ length_1_m       : num [1:5] 0 0 0.493 1.145 2.086
#>   ..$ branches         : int [1:5] 0 5 30 22 14
#>   ..$ branches_1       : int [1:5] 0 0 3 3 4
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_height_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	4 obs. of  9 variables:
#>   ..$ height_class_m: num [1:4] 1 2 3 4
#>   ..$ volume_m3     : num [1:4] 0 0.00434 0.00512 0.00192
#>   ..$ volume_1_m3   : num [1:4] 0 0.003577 0.0018 0.000419
#>   ..$ area_m2       : num [1:4] 0 0.622 0.856 0.404
#>   ..$ area_1_m2     : num [1:4] 0 0.472 0.2706 0.0847
#>   ..$ length_m      : num [1:4] 0 8.27 13.61 7.46
#>   ..$ length_1_m    : num [1:4] 0 5.72 3.7 1.42
#>   ..$ branches      : int [1:4] 0 14 36 31
#>   ..$ branches_1    : int [1:4] 0 5 7 2
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_angle_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	18 obs. of  9 variables:
#>   ..$ angle_class_deg: num [1:18] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3      : num [1:18] 0.00 6.27e-05 1.30e-04 6.26e-04 2.28e-03 ...
#>   ..$ volume_1_m3    : num [1:18] 0 0 0 0.000518 0.001035 ...
#>   ..$ area_m2        : num [1:18] 0 0.0145 0.0272 0.1101 0.3457 ...
#>   ..$ area_1_m2      : num [1:18] 0 0 0 0.0833 0.1651 ...
#>   ..$ length_m       : num [1:18] 0 0.27 0.496 1.775 4.619 ...
#>   ..$ length_1_m     : num [1:18] 0 0 0 1.18 2.24 ...
#>   ..$ branches       : int [1:18] 0 1 2 4 8 14 15 17 11 4 ...
#>   ..$ branches_1     : int [1:18] 0 0 0 1 2 4 3 1 1 2 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_zenith_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	18 obs. of  9 variables:
#>   ..$ zenith_class_deg: num [1:18] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3       : num [1:18] 3.72e-05 2.53e-04 2.19e-03 4.86e-04 3.21e-03 ...
#>   ..$ volume_1_m3     : num [1:18] 0 0 0.0021 0.00013 0.00192 ...
#>   ..$ area_m2         : num [1:18] 0.00858 0.04931 0.25267 0.1042 0.45743 ...
#>   ..$ area_1_m2       : num [1:18] 0 0 0.2359 0.0291 0.2698 ...
#>   ..$ length_m        : num [1:18] 0.16 0.855 2.676 2.001 5.832 ...
#>   ..$ length_1_m      : num [1:18] 0 0 2.399 0.527 3.361 ...
#>   ..$ branches        : int [1:18] 1 3 3 10 8 14 14 10 8 5 ...
#>   ..$ branches_1      : int [1:18] 0 0 1 1 4 1 3 2 1 1 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_azimuth_dist      :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	36 obs. of  9 variables:
#>   ..$ azimuth_class_deg: num [1:36] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3        : num [1:36] 4.05e-04 9.39e-05 4.42e-06 2.69e-05 3.36e-05 ...
#>   ..$ volume_1_m3      : num [1:36] 2.88e-04 0.00 0.00 0.00 1.38e-05 ...
#>   ..$ area_m2          : num [1:36] 0.08238 0.01788 0.00219 0.00932 0.01128 ...
#>   ..$ area_1_m2        : num [1:36] 0.05563 0 0 0 0.00503 ...
#>   ..$ length_m         : num [1:36] 1.4416 0.3695 0.0888 0.2682 0.3045 ...
#>   ..$ length_1_m       : num [1:36] 0.888 0 0 0 0.147 ...
#>   ..$ branches         : int [1:36] 4 3 1 2 2 2 2 3 1 1 ...
#>   ..$ branches_1       : int [1:36] 1 0 0 0 1 0 0 0 0 1 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_order_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	5 obs. of  5 variables:
#>   ..$ branch_order: int [1:5] 1 2 3 4 5
#>   ..$ branches    : int [1:5] 14 35 23 8 1
#>   ..$ volume_m3   : num [1:5] 5.80e-03 3.86e-03 1.06e-03 6.22e-04 3.46e-05
#>   ..$ area_m2     : num [1:5] 0.82732 0.73184 0.2375 0.07846 0.00693
#>   ..$ length_m    : num [1:5] 10.83 12.3 4.66 1.44 0.11
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_reverse_order_dist:Classes ‘tidytable’, ‘data.table’ and 'data.frame':	8 obs. of  5 variables:
#>   ..$ reverse_order: num [1:8] 1 2 3 4 5 6 7 16
#>   ..$ segments     : int [1:8] 52 16 4 2 2 2 2 1
#>   ..$ volume_m3    : num [1:8] 0.002042 0.001876 0.000557 0.000648 0.00051 ...
#>   ..$ area_m2      : num [1:8] 0.459 0.348 0.111 0.112 0.105 ...
#>   ..$ length_m     : num [1:8] 9.22 6.08 1.88 1.7 1.82 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ stem_taper               :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	88 obs. of  2 variables:
#>   ..$ height_m   : num [1:88] 0 0.0939 0.1661 0.2326 0.242 ...
#>   ..$ diameter_cm: num [1:88] 9.29 9.08 8.84 8.74 8.59 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ spreads                  :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	360 obs. of  3 variables:
#>   ..$ azimuth_deg : num [1:360] 20 40 60 80 100 120 140 160 180 200 ...
#>   ..$ height_class: int [1:360] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ spread_m    : num [1:360] 0.0933 0.0938 0.0908 0.0955 0.0962 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ vertical_profile         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	20 obs. of  2 variables:
#>   ..$ height_class: int [1:20] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ avg_spread_m: num [1:20] 0.0953 0.0892 0.0893 0.0802 0.0845 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
#> Updating Branch Ordering
#> Calculating Total Children
#> Building Cylinder Network
#> Calculating Distance From Base
#> Calculating Average Distance To Twigs
metrics <- tree_metrics(cylinder)
#> Calculating Branch Metrics
#> Calculating Tree Metrics
#> Generating Point Cloud
#> Calculating Crown Metrics
#> Calculating Tree Distributions
#> Calculating Branch Distributions
#> Calculating Stem Taper
#> Calculating Spreads
str(metrics)
#> List of 17
#>  $ branch                   :Classes ‘grouped_tt’, ‘tidytable’, ‘data.table’ and 'data.frame':	65 obs. of  19 variables:
#>   ..$ branch          : int [1:65] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ parent_branch   : int [1:65] 0 1 1 1 1 5 1 1 1 9 ...
#>   ..$ branch_order    : int [1:65] 0 1 1 1 1 2 1 1 1 2 ...
#>   ..$ reverse_order   : int [1:65] 18 1 1 1 2 1 1 1 4 1 ...
#>   ..$ diameter_base_cm: num [1:65] 9.44 1.52 1.96 2.87 2.37 ...
#>   ..$ diameter_mid_cm : num [1:65] 2.387 0.815 1.5 1.239 0.85 ...
#>   ..$ diameter_tip_cm : num [1:65] 1.12 1.26 1.16 1.72 1.47 ...
#>   ..$ volume_m3       : num [1:65] 1.29e-02 2.69e-05 7.65e-05 3.07e-04 1.96e-04 ...
#>   ..$ area_m2         : num [1:65] 0.76808 0.00716 0.01581 0.04765 0.03439 ...
#>   ..$ length_m        : num [1:65] 4.321 0.153 0.274 0.615 0.501 ...
#>   ..$ height_m        : num [1:65] 0 3.23 3.19 3.17 3.16 ...
#>   ..$ angle_deg       : num [1:65] 0 46.5 61.5 81.8 86.6 ...
#>   ..$ azimuth_deg     : num [1:65] -109.9 126.8 165 -164.2 -30.7 ...
#>   ..$ zenith_deg      : num [1:65] 4.82 81.98 19.69 95.58 64.21 ...
#>   ..$ growth_length   : num [1:65] 31.4264 0.0939 0.2738 0.615 0.5729 ...
#>   ..$ nodes           : int [1:65] 18 2 1 1 2 1 1 1 4 1 ...
#>   ..$ children        : num [1:65] 18 4 1 1 2 1 1 1 4 1 ...
#>   ..$ base_distance_m : num [1:65] 0.0497 4.2489 4.0634 3.7686 3.6042 ...
#>   ..$ twig_distance_m : num [1:65] 3.4286 0.0939 0.2738 0.615 0.4527 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..- attr(*, "groups")= chr "branch"
#>  $ cloud                    : num [1:12262, 1:3] 0.787 0.796 0.803 0.807 0.807 ...
#>  $ tree                     :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	1 obs. of  27 variables:
#>   ..$ tree_volume_m3         : num 0.03
#>   ..$ stem_volume_m3         : num 0.0129
#>   ..$ branch_volume_m3       : num 0.0171
#>   ..$ tree_height_m          : num 3.67
#>   ..$ tree_length_m          : num 31.4
#>   ..$ stem_length_m          : num 4.32
#>   ..$ branch_length_m        : num 27.1
#>   ..$ branches               : int 65
#>   ..$ max_branch_order       : int 4
#>   ..$ max_reverse_order      : int 18
#>   ..$ tree_area_m2           : num 3.03
#>   ..$ stem_area_m2           : num 0.768
#>   ..$ branch_area_m2         : num 2.26
#>   ..$ dbh_qsm_cm             : num 7.39
#>   ..$ dbh_raw_cm             : num 7.39
#>   ..$ crown_diameter_mean_m  : num 1.52
#>   ..$ crown_diameter_max_m   : num 2.3
#>   ..$ crown_diameter_min_m   : num 0.89
#>   ..$ crown_projected_area_m2: num 3.28
#>   ..$ crown_base_height_m    : num 2.27
#>   ..$ crown_length_m         : num 1.4
#>   ..$ crown_ratio            : num 0.381
#>   ..$ crown_area_m2          : num 9.93
#>   ..$ crown_volume_m3        : num 0.00234
#>   ..$ start_x                : num -0.0286
#>   ..$ start_y                : num -0.0789
#>   ..$ start_z                : num 0.996
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_height_dist         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	4 obs. of  4 variables:
#>   ..$ height_class_m: int [1:4] 1 2 3 4
#>   ..$ volume_m3     : num [1:4] 0.0059 0.0072 0.01098 0.00589
#>   ..$ area_m2       : num [1:4] 0.273 0.563 1.319 0.872
#>   ..$ length_m      : num [1:4] 1 4.53 14.56 11.33
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_diameter_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	10 obs. of  4 variables:
#>   ..$ diameter_class_cm: num [1:10] 1 2 3 4 5 6 7 8 9 10
#>   ..$ volume_m3        : num [1:10] 1.17e-05 1.66e-03 6.23e-03 5.64e-03 2.58e-03 ...
#>   ..$ area_m2          : num [1:10] 0.00478 0.38688 1.00109 0.64122 0.23681 ...
#>   ..$ length_m         : num [1:10] 0.156 7.332 12.956 5.841 1.736 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_zenith_dist         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	14 obs. of  4 variables:
#>   ..$ zenith_class_deg: num [1:14] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3       : num [1:14] 0.0089 0.00409 0.00348 0.00237 0.00305 ...
#>   ..$ area_m2         : num [1:14] 0.479 0.342 0.345 0.303 0.38 ...
#>   ..$ length_m        : num [1:14] 2.34 2.88 3.25 3.63 4.35 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ tree_azimuth_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	36 obs. of  4 variables:
#>   ..$ azimuth_class_deg: num [1:36] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3        : num [1:36] 0.000517 0.000698 0.000605 0.000394 0.000213 ...
#>   ..$ area_m2          : num [1:36] 0.0656 0.0623 0.0624 0.0548 0.0318 ...
#>   ..$ length_m         : num [1:36] 0.789 0.638 0.725 0.71 0.416 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_diameter_dist     :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	6 obs. of  9 variables:
#>   ..$ diameter_class_cm: num [1:6] 1 2 3 4 5 6
#>   ..$ volume_m3        : num [1:6] 0 0 0.000108 0.001386 0.001816 ...
#>   ..$ volume_1_m3      : num [1:6] 0.00 0.00 2.82e-05 1.03e-04 6.95e-04 ...
#>   ..$ area_m2          : num [1:6] 0 0 0.0243 0.2515 0.3099 ...
#>   ..$ area_1_m2        : num [1:6] 0 0 0.00727 0.02297 0.12553 ...
#>   ..$ length_m         : num [1:6] 0 0 0.455 3.935 4.493 ...
#>   ..$ length_1_m       : num [1:6] 0 0 0.15 0.427 1.89 ...
#>   ..$ branches         : int [1:6] 0 0 3 19 14 15
#>   ..$ branches_1       : int [1:6] 0 0 1 2 5 4
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_height_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	4 obs. of  9 variables:
#>   ..$ height_class_m: num [1:4] 1 2 3 4
#>   ..$ volume_m3     : num [1:4] 0 0.0065 0.00808 0.00248
#>   ..$ volume_1_m3   : num [1:4] 0 0.005473 0.002528 0.000695
#>   ..$ area_m2       : num [1:4] 0 0.772 1.078 0.407
#>   ..$ area_1_m2     : num [1:4] 0 0.602 0.354 0.118
#>   ..$ length_m      : num [1:4] 0 8.36 12.96 5.78
#>   ..$ length_1_m    : num [1:4] 0 5.91 4.39 1.7
#>   ..$ branches      : int [1:4] 0 13 29 22
#>   ..$ branches_1    : int [1:4] 0 6 6 5
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_angle_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	18 obs. of  9 variables:
#>   ..$ angle_class_deg: num [1:18] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3      : num [1:18] 0 0 0.000158 0.001303 0.005613 ...
#>   ..$ volume_1_m3    : num [1:18] 0.0 0.0 0.0 8.9e-05 4.6e-03 ...
#>   ..$ area_m2        : num [1:18] 0 0 0.0248 0.1747 0.6259 ...
#>   ..$ area_1_m2      : num [1:18] 0 0 0 0.0134 0.4767 ...
#>   ..$ length_m       : num [1:18] 0 0 0.317 2.059 6.412 ...
#>   ..$ length_1_m     : num [1:18] 0 0 0 0.16 4.45 ...
#>   ..$ branches       : int [1:18] 0 0 2 5 11 7 9 16 9 3 ...
#>   ..$ branches_1     : int [1:18] 0 0 0 1 3 0 2 3 7 1 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_zenith_dist       :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	18 obs. of  9 variables:
#>   ..$ zenith_class_deg: num [1:18] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3       : num [1:18] 0 0.000479 0.001105 0.000417 0.003383 ...
#>   ..$ volume_1_m3     : num [1:18] 0.00 7.65e-05 8.90e-05 0.00 3.13e-03 ...
#>   ..$ area_m2         : num [1:18] 0 0.079 0.1425 0.0719 0.3334 ...
#>   ..$ area_1_m2       : num [1:18] 0 0.0158 0.0134 0 0.2973 ...
#>   ..$ length_m        : num [1:18] 0 1.13 1.61 1.04 2.95 ...
#>   ..$ length_1_m      : num [1:18] 0 0.274 0.16 0 2.513 ...
#>   ..$ branches        : int [1:18] 0 4 4 6 2 6 10 8 4 8 ...
#>   ..$ branches_1      : int [1:18] 0 1 1 0 1 1 3 0 2 5 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_azimuth_dist      :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	36 obs. of  9 variables:
#>   ..$ azimuth_class_deg: num [1:36] 10 20 30 40 50 60 70 80 90 100 ...
#>   ..$ volume_m3        : num [1:36] 5.24e-05 3.07e-04 8.39e-05 1.52e-04 4.57e-05 ...
#>   ..$ volume_1_m3      : num [1:36] 0 0.000307 0 0 0 ...
#>   ..$ area_m2          : num [1:36] 0.0108 0.04765 0.0181 0.02557 0.00883 ...
#>   ..$ area_1_m2        : num [1:36] 0 0.0476 0 0 0 ...
#>   ..$ length_m         : num [1:36] 0.188 0.615 0.324 0.365 0.146 ...
#>   ..$ length_1_m       : num [1:36] 0 0.615 0 0 0 ...
#>   ..$ branches         : int [1:36] 1 1 3 2 1 0 2 2 0 0 ...
#>   ..$ branches_1       : int [1:36] 0 1 0 0 0 0 1 1 0 0 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_order_dist        :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	4 obs. of  5 variables:
#>   ..$ branch_order: int [1:4] 1 2 3 4
#>   ..$ branches    : int [1:4] 17 26 18 3
#>   ..$ volume_m3   : num [1:4] 8.70e-03 6.46e-03 1.82e-03 7.23e-05
#>   ..$ area_m2     : num [1:4] 1.0741 0.8713 0.2976 0.0148
#>   ..$ length_m    : num [1:4] 12.01 10.63 4.216 0.249
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ branch_reverse_order_dist:Classes ‘tidytable’, ‘data.table’ and 'data.frame':	8 obs. of  5 variables:
#>   ..$ reverse_order: int [1:8] 1 2 3 4 5 6 8 11
#>   ..$ segments     : int [1:8] 47 8 2 3 1 1 1 1
#>   ..$ volume_m3    : num [1:8] 0.004292 0.001983 0.000571 0.002081 0.001349 ...
#>   ..$ area_m2      : num [1:8] 0.715 0.313 0.091 0.27 0.171 ...
#>   ..$ length_m     : num [1:8] 10.36 4.23 1.22 3.08 1.82 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ stem_taper               :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	123 obs. of  2 variables:
#>   ..$ height_m   : num [1:123] 0 0.0497 0.1027 0.1562 0.2086 ...
#>   ..$ diameter_cm: num [1:123] 9.44 9.58 9.38 9.33 9.07 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ spreads                  :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	360 obs. of  3 variables:
#>   ..$ azimuth_deg : num [1:360] 20 40 60 80 100 120 140 160 180 200 ...
#>   ..$ height_class: int [1:360] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ spread_m    : num [1:360] 0.1055 0.0994 0.1075 0.1075 0.1081 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ vertical_profile         :Classes ‘tidytable’, ‘data.table’ and 'data.frame':	20 obs. of  2 variables:
#>   ..$ height_class: int [1:20] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ avg_spread_m: num [1:20] 0.1015 0.0956 0.0917 0.0862 0.0875 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr>