Plot QSM

Efficiently plot QSMs and point clouds. Uses the Rcpp and RGL libraries as backends.

Usage

plot_qsm(
  cylinder = NULL,
  radius = NULL,
  color = NULL,
  palette = NULL,
  facets = 6,
  skeleton = FALSE,
  skeleton_lwd = NULL,
  cloud = NULL,
  pt_color = NULL,
  pt_size = NULL,
  triangulation = NULL,
  tri_color = NULL,
  tri_palette = NULL,
  axes = TRUE,
  axes_color = NULL,
  hover = FALSE,
  bg_color = NULL,
  lit = TRUE,
  pan = TRUE,
  normalize = FALSE
)

Arguments

cylinder

QSM cylinder data frame

radius

Column name of radii as a quoted string. Defaults to the modified radii.

color

Optional cylinder color parameter. Colors must be a single hex color, a vector of hex colors, or a quoted column name. It can also be set to "random" to generate a random solid color. Vectors must have the same length as the cylinder data frame.

palette

Optional color palette for numerical data. Palettes include colourvalues::color_palettes() or any user supplied palette.

facets

The number of facets in the polygon cross section. Defaults to 6, but can be increased to improve visual smoothness at the cost of performance and memory.

skeleton

Plot the QSM skeleton instead of cylinders. Defaults to FALSE.

skeleton_lwd

Skeleton line width. Defaults to 1.

cloud

Point cloud data frame where the first three columns are the x, y, and z coordinates in the same coordinate system as the QSM. Defaults to NULL.

pt_color

Color of the point cloud. Defaults to black. Can be set to "random".

pt_size

Size of the points. Defaults to 0.1.

triangulation

Plot the stem triangulation mesh from TreeQSM. Defaults to NULL

tri_color

Color of the triangulation mesh. Colors must be a single hex color.

tri_palette

Optional triangulation color palette for z values. Supports the same inputs as palettes.

axes

Show plot axes. Defaults to TRUE.

axes_color

Set the axes color. Defaults to black.

hover

Show cylinder and branch id on mouse hover. Defaults to FALSE.

bg_color

Set the background color of the plot. Defaults to white.

lit

Enable light source in plot. Defaults to TRUE. Can be set to FALSE.

pan

Use right mouse button to pan plot. Defaults to TRUE, but is disabled when hover is enabled.

normalize

Normalize the base of the QSM to 0,0,0. Defaults to FALSE.

Value

A rgl plot

Examples

## TreeQSM Processing Chain & Triangulation
file <- system.file("extdata/QSM.mat", package = "rTwig")
qsm <- import_qsm(file)
#> Importing TreeQSM .mat
cylinder <- qsm$cylinder
cylinder <- update_cylinders(cylinder)
#> Updating Cylinder Ordering
#> Calculating Total Children
#> Building Cylinder Network
#> Verifying Topology
#> Calculating Growth Length
#> Calculating Reverse Branch Order
#> Calculating Branch Segments
#> Calculating Alternate Branch Numbers
#> Calculating Distance From Base
#> Calculating Average Distance To Twigs
plot_qsm(cylinder)
#> Plotting Cylinders
#> Panning plot on rgl device: 1

triangulation <- qsm$triangulation
plot_qsm(triangulation = triangulation)
#> Plotting Triangulation
#> Panning plot on rgl device: 1

## SimpleForest Processing Chain
file <- system.file("extdata/QSM.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
#> Finding Branches
#> Calculating Total Children
#> Building Cylinder Network
#> Verifying Topology
#> Calculating Distance From Base
#> Calculating Average Distance To Twigs
plot_qsm(cylinder)
#> Plotting Cylinders
#> Panning plot on rgl device: 1

## aRchi Processing Chain
file <- system.file("extdata/QSM2.csv", package = "rTwig")
cylinder <- read.csv(file)
cylinder <- update_cylinders(cylinder)
#> Updating Cylinder Ordering
#> Finding Branches
#> Calculating Total Children
#> Building Cylinder Network
#> Verifying Topology
#> Calculating Growth Length
#> Calculating Reverse Branch Order
#> Calculating Branch Segments
#> Calculating Alternate Branch Numbers
#> Calculating Distance From Base
#> Calculating Average Distance To Twigs
plot_qsm(cylinder)
#> Plotting Cylinders
#> Panning plot on rgl device: 1